Dynamically create JSON representation from vars
Currently, I'm relying on a special case to elimninate "target" from
being including in this representation. This might indicate a few
different things-
1. It should be a hidden variable, and I should adjust my other code
that access this field accordingly.
2. I shouldn't have a persistant target as part of a Unit at all. I
could just determine the target in each loop and handle it there,
without it being tied in with the attacking Unit.
3. Let it get included with the other fields as if it's no different. I
don't think this is a good idea, because if a game's objects get
serialized and reloaded later, they're not going to have the same
references, making this field useless at best, misleading at worst.