Bring PriorityQueue into algorithms
I'm not truly making use of this yet, because I need to figure out how
to retrieve edge weights from the graph via the stop/start vertices.

Also, I'm passing in the current time as a fallback sorting option for
spots in the frontier. It might be worthwhile to add a comparison
operator for the BoardSpot class so that every insertion into a
PriorityQueue doesn't need that extra param.